moved clickhouse migrations and stuff
