+++ tensorzero-internal/src/inference/providers/fireworks.rs
@@ -488,12 +488,12 @@ mod tests {
 
     use super::*;
 
-    use crate::inference::providers::common::{WEATHER_TOOL, WEATHER_TOOL_CONFIG};
     use crate::inference::providers::openai::{
         OpenAIFinishReason, OpenAIResponseChoice, OpenAIResponseMessage, OpenAIToolType,
         OpenAIUsage,
     };
     use crate::inference::providers::openai::{SpecificToolChoice, SpecificToolFunction};
+    use crate::inference::providers::test_helpers::{WEATHER_TOOL, WEATHER_TOOL_CONFIG};
     use crate::inference::types::{FunctionType, RequestMessage, Role};
 
     #[test]
