+++ tensorzero-internal/src/inference/providers/gcp_vertex_gemini.rs
@@ -1198,7 +1198,7 @@ mod tests {
     use serde_json::json;
 
     use super::*;
-    use crate::inference::providers::common::{MULTI_TOOL_CONFIG, QUERY_TOOL, WEATHER_TOOL};
+    use crate::inference::providers::test_helpers::{MULTI_TOOL_CONFIG, QUERY_TOOL, WEATHER_TOOL};
     use crate::inference::types::{FunctionType, ModelInferenceRequestJsonMode};
     use crate::tool::{ToolCallConfig, ToolResult};
