+++ tensorzero-internal/src/inference/providers/mod.rs
@@ -1,8 +1,6 @@
 pub mod anthropic;
 pub mod aws_bedrock;
 pub mod azure;
-#[cfg(test)]
-pub mod common;
 pub mod deepseek;
 #[cfg(any(test, feature = "e2e_tests"))]
 pub mod dummy;
@@ -16,6 +14,8 @@ pub mod mistral;
 pub mod openai;
 pub mod provider_trait;
 pub mod sglang;
+#[cfg(test)]
+pub mod test_helpers;
 pub mod tgi;
 pub mod together;
 pub mod vllm;
