+++ tensorzero-internal/src/inference/providers/sglang.rs
@@ -461,10 +461,10 @@ mod tests {
 
     use crate::inference::{
         providers::{
-            common::WEATHER_TOOL_CONFIG,
             openai::{
                 OpenAIFinishReason, OpenAIResponseChoice, OpenAIResponseMessage, OpenAIUsage,
             },
+            test_helpers::WEATHER_TOOL_CONFIG,
         },
         types::{FinishReason, FunctionType, ModelInferenceRequestJsonMode, RequestMessage, Role},
     };
