+++ tensorzero-internal/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod cache;
 pub mod clickhouse;
-pub mod clickhouse_migration_manager;
 pub mod config_parser; // TensorZero config file
 pub mod embeddings; // embedding inference
 pub mod endpoints; // API endpoints
